# api/kube-jit/docs/swagger.yaml
basePath: /kube-jit-api
definitions:
...
  models.SimpleMessageResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Self-service Kubernetes RBAC JIT Requests with Google/Azure/Github
    Oauth.
  title: Kube-JIT API
  version: "1.0"
paths:
...
  /kube-jit-api/approve-reject:
    post:
      consumes:
      - application/json
      description: |-
        Approves or rejects pending JIT access requests. Admins and platform approvers can approve/reject multiple requests at once. Non-admins can approve/reject individual namespaces.
        Requires one or more cookies named kube_jit_session_<number> (e.g., kube_jit_session_0, kube_jit_session_1).
        Pass split cookies in the Cookie header, for example:
        -H "Cookie: kube_jit_session_0=${cookie_0};kube_jit_session_1=${cookie_1}"
        Note: Swagger UI cannot send custom Cookie headers due to browser security restrictions. Use curl for testing with split cookies.
      parameters:
      - description: 'Session cookies (multiple allowed, names: kube_jit_session_0,
          kube_jit_session_1, etc.)'
        in: header
        name: Cookie
        required: true
        type: string
      - description: Approval/rejection payload (structure depends on user role)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Requests processed successfully
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "401":
          description: 'Unauthorized: no approver groups in session'
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Failed to process requests
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Approve or reject JIT access requests
      tags:
      - request
...
  /kube-jit-api/submit-request:
    post:
      consumes:
      - application/json
      description: |-
        Creates a new JIT access request for the authenticated user.
        Requires one or more cookies named kube_jit_session_<number> (e.g., kube_jit_session_0, kube_jit_session_1).
        Pass split cookies in the Cookie header, for example:
        -H "Cookie: kube_jit_session_0=${cookie_0};kube_jit_session_1=${cookie_1}"
        Note: Swagger UI cannot send custom Cookie headers due to browser security restrictions. Use curl for testing with split cookies.
      parameters:
      - description: 'Session cookies (multiple allowed, names: kube_jit_session_0,
          kube_jit_session_1, etc.)'
        in: header
        name: Cookie
        required: true
        type: string
      - description: JIT request payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Request submitted successfully
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "401":
          description: 'Unauthorized: no token in session data'
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Failed to submit request
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Submit a new JIT access request
      tags:
      - request
swagger: "2.0"